requirements.txt
maikol-utils>=0.8.0
dotenv>=0.9.9
setuptools>=80.9.0
PyYAML>=6.0
